refactor CUDA code into its own sets of files Summary:
Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
